Refactor binary extraction for macOS
Currently the binary for hadolint on their GitHub page is segfaulting.
Since this will likely be addressed in the future I am making it so the
binaries are extracted from a Homebrew bottle instead. The checksum has
and the URL are hardcoded so it might need to be updated if the
segfaulting is still an issue in the next release.